Split time zone scripts: download and update
Split time zone scripts to produce separate download and update
scripts. An initial copy of IANA's tzdata2017b file is checked
in here (this matches the current version of the data found in
bionic / external/icu).

Added copyright headers and a license / notice file for
the IANA files.

Check in a copy of bionic/libc/zoneinfo/tzlookup.xml to serve
as an "input file". The update-tzdata.py copies it exactly to
bionic as an "output" file.

There will be a follow-up change to move the output files
(tzdata and tzlookup.xml) from bionic to an "output data"
directory, removing the need to reference bionic for time
zone data files at all.

Bug: 31008728
Test: Deleted tzdata2017b.tar.gz, ran ./download-iana-data.py
Test: Ran ./update-tzdata.py and checked the output
Test: diff \
    $ANDROID_BUILD_TOP/system/timezone/input_data/android/tzlookup.xml \
    $ANDROID_BUILD_TOP/bionic/libc/zoneinfo/tzlookup.xml

Change-Id: I2d3e42a6c3efe10f2340c042f29462720443aa4a