Increment the distro major release version
Increment the distro major release version
as a result of the update to ICU59.

This contains a fix for
TimeZoneDistroInstallerTest which was
accidentally assuming the major version
was < 2 along with a small refactoring
to reduce the number of occurrences of
major/minor version numbers in code.

This change contains the documentation updates
in README.android that describe the process
followed (including testing).

Test: See README.android
Bug: 64712155
Change-Id: If9f32c30f0c20956c847a9f81e492db128674f3f