Generate the tzlookup.xml file from a proto file
This commit adds a platform developer command-line
tool to generate the tzlookup.xml from a human-readable
.proto file rather than just copying the .xml file
around.

The only files that influence on-device behavior
are those in system/timezone/output_data.

The new countryzones.txt is a translation of the
old tzlookup.xml with more structure and some
additional data. A few offsets were found to be
incorrect by the new validation code and these have
been corrected (this information was previously found
in comments in tzlookup.xml and was never actually used
on device).

During generation the tzlookup_generator can validate
the input against IANA and ICU data and do some
defaulting which should help avoid the data becoming
out of date.

This change introduces a country-level attribute to the
tzlookup.xml for the "country default" time zone, which is
there to help address bug 34266809 in a follow-up
change.

This change also moves the zone_compactor tool to
being built with the Android build system for consistency.

The change updates the update-tzdata.py script to use the
new tool.

The updated output_data files from running the tool are
included in this commit.

The new tool has tests instructions for execution are
in README.android.

Bug: 34266809
Bug: 65575839
Test: Booted device
Test: See included host tests.
Change-Id: Ic96af7987c61690faebbc80907209f9e1cc1e976