Modify the default time zones for regions
The default time zone for each region in tzlookup.xml
is new and was initially set using the first time zone
ID given in the list of time zones for the region in
the old input_data/android/tzlookup.xml file.

The default values have now been compared with the first
zone ID given in each region in zone.tab from IANA (from
2017b), which is what Android's MccTable class has
historically used to generate the default for a region
when only the ISO code was available.

Where there was a difference between zone.tab and
tzlookup.xml these have been investigated and a new
explicit default chosen where needed, usually
the time zone containing the capital for the region.

This is being done in preparation for changing MccTable
over to using the default value in tzlookup.xml instead of
zone.tab: with this change the default chosen won't be
worse and Android will have explicit control over the
default without needing special-cases in the code.

Bug: 36971201
Bug: 34266809
Test: Inspection of the generated tzlookup.xml
Change-Id: Ice67330b3904c62345e28151f5f7855b2f8f53f3