Update platform files to 2017c
This change updates OC-MR1 to 2017c. It contains a local
patch to IANA's tzdata2017c.tar.gz file to modify the backzone
file and remove the Africa/Juba entry (which was a
near-duplicate of the one in the africa file and causes
problems for the zic/tzdata generation script). The patch
originates from:
http://mm.icann.org/pipermail/tz/2017-October/025380.html
and has been included in the tzdata2017c.tar.gz file.

The output_data files (and updated data files in
external/icu) were generated using:
system/timezone/update-tzdata.py

There is an associated change in external/icu containing
upstream ICU metadata changes and the generated .dat files.

Bug: 68136449
Test: run cts -m CtsLibcoreTestCases
Test: run cts -m CtsBionicTestCases
Test: run cts -m CtsIcuTestCases
Change-Id: I9ef3cbd80e94c0bc7ba544091813e8f31a1af536