Add missing default= to test XML
Add missing country default attribute to XML used
in TimeZoneDistroInstallerTest. The information is
now required by the TimeZoneFinder and the
TimeZoneDistroInstaller will fail to validate the XML
without it.

This change also adds the ianaversion attribute to the
top of the XML. It's not used at runtime but is
better representative of real data.

Now that tzlookup.xml is generated on the host it would
be ideal if the test XML were generated in code and not
a hardcoded String but that would require access to
various JAXP APIs not currently present on Android.

Bug: 36971201
Test: CTS: run cts -m CtsLibcoreTestCases -t com.android.timezone.distro.installer.TimeZoneDistroInstallerTest
Change-Id: I7d72d87eb5153af90b15a6cc0e404606f9bb17eb