Improve tz update scripts
The update-tzdata.py script has historically used
the zic binary in the user's PATH. This change moves to
building and using upstream IANA code. The version
initially chosen (2014b) is deliberate: it is the
newest version that does not alter the files already
submitted and were generated by the zic version
historically used by the AOSP developers.

This change also makes some improvements to the
download-iana-data.py tool to record the
signature file associated with the data.

Bug: 29383204
Test: Manual execution / testing
Change-Id: I99d374b8efb849a23e373e477237a6ac54de87b0