Annotating zones that are not strictly necessary.
This change adds a host-side code to help reduce the
number of zones considered for a new time zone picker.
The TzLookupGenerator will now output additional
"notafter" attributes for zones that do not need
to be considered after a point in time.

The tool relies on data from ICU to establish zone
rules / naming information and the priority
information held in countryzones.txt. When two or
more zones are effectively identical for a period
the priority is used to decide which zone should
carry forward and which zone should have
"notafter" added.

Also included is a tool (UniqueZonesVisualizer)
to help debugging problems with zone prioritization
with tools like "dot".

This change contains an integration test
for TzLookupGenerator. zonetree classes are
currently lacking unit tests. This will be resolved
in a follow up commit when time allows.

Bug: 72142943
Test: Visual inspection of tzlookup.xml
Test: Ran tzlookup_generator/ tests, see README.android
Change-Id: I37568bc923bcaadb8b1305eec5dc1fea2ae3012b