More tweaks to tzlookup.xml
This change is to allow for more meaningful
priorities for zones in countryzones.txt.

The exact meaning of priority is left abstract
since different countries have different
zone naming schemes assigned to them.

The priorities were orginally intended to
be abstract and just to influence priorities
between individual nodes in subtrees but this
is going to be difficult to maintain and error
prone for large countries. For large countries it
will make sense to relate priority to some
metric associated with the zone, usually
population. This has now been done for the
larger countries. This highlighted a few
issues with Brazil where the city population
had incorrectly been used instead of a state
population when previously choosing priority.

The only change users would see is that Brazil
will now show Ciuaba instead of Campo Grande.

Included are some improvements to the host code:
improvements to the visualization debugging tool
to add more information to the graphviz output
and removal of unused imports.

As always, files are generated by update-tzdata.py
in system/timezone.

Test: make droid / make tzlookup_generator / inspection
Bug: 77227594
Change-Id: I1836f908ee0a67406731bb01525ffde7c9448838