Improvements to host-side tools
This commit contains some comment fixes suggested by
a reviewer on a cherry-pick.

It also corrects some behavior in CountryZoneTree when the
priorities haven't been set correctly: if there were a priority
clash the graph-viz generation would fail with an error
because getPrimaryZoneInfo() requires the priorities
have been resolved. This has been replaced with just getting
an arbitrary zoneinfo.

The change also improves TzLookupGenerator behavior when
there are priority clashes for multiple countries: errors
are reported for every country, not just the first, as
was originally intended but had been lost with some other
recent changes.

Test: Ran the unit tests
Test: Ran update-tzdata.py and verified no data file changes
Bug: 77227594
Change-Id: I0999b4106ff63070bbef95f05524eaaa3b0f36ac