Update the zic version used when generating tzdata
Historically, Android has used zic 2014b (or before)
to generate the tzdata file used by libcore's ZoneInfo
and bionic's mktime / localtime, etc.

Originally the version used was implicit and was based
on the zic version available on the developer's machine.
The version to be used was made explicit for consistency
in commit cdf5695c2b83e2781c287d8d315eb903554940da.

This change upgrades the tools version to the most
recent available at the time of writing and also modifies
/ renames the download-iana-data.py script so it can also
download the files needed for building zic.

There are associated changes in platform/libcore to handle
the test changes associated with the zic upgrade.

Bug: 73719425
Test: Manual execution of download-iana-files.py with missing and
old files
Test: run cts -m CtsLibcoreTestCases
Test: run cts -m CtsBionicTestCases
Change-Id: Ic11af7e586e4dc3597745dcaccf57cf8cb6e082a