Explicitly include version info in output_data
Explicitly include a version file in /system/timezone/output_data
so that we can use it elsewhere. The version file contains information
about the format of the content and the content version (e.g. 2018e).

The version file is used in the "Time zone updates via APK" feature
to check / enforce compatibility between time zone distro.zip files and
code (e.g. ICU) on the device.

Previously the file was only included inside the distro.zip file.
Including it outside of the distro.zip avoids us needing to duplicate
version information for the APEX build or come up with a new versioning
strategy. For a while at least distro.zip files will co-exist alongside
APEX files and they have very similar versioning requirements.

Versioning concerns are mostly going to be handled by the
distribution mechanism for APEX files so the APEX files
should always match the device expectations.

That said, it will be good to have an indicator of compatibility / the
dessert release the APEX was generated for. This could be used by
developers when looking at image files before install and for debugging
the state on device if (somehow) an APEX file has been installed on
an incompatible device. Checking the version on device could be
incorporated into a pre-install check later if necessary.

Bug: 115612964
Test: cd system/timezone; ./update-tzdata.py [then inspect output]
Change-Id: I405deaea27942686801ba6052dc47c4c2969d764