Refactor DistroVersion, create TzDataSetVersion
DistroVersion was created for the time zone updates via APK
feature to handle data and format versioning.

This change refactors DistroVersion. It still defines the
format of the distro_version file inside of a time zone data
distro.zip.

A new class, TzDataSetVersion is being added to libcore to
define the format of the tz_version file used in the newer
time zone data module. The knowledge of the format version
used on the device now moves there. The TzDataSetVersion is
also being exposed in a host library for host-side tests.

Bug: 119026403
Test: CTS: run cts-dev -m CtsLibcoreTestCases
Change-Id: I9d9b999336bdd1185a5014a866a508371733d782