Changes to support RuntimeService in framework
Revert TimeZoneDistroInstaller to using DistroVersion
in one place. This avoids making some TzDataSetVersion
fields public.

This commit also adds build target to include the tz_version
file in the system image for debug usage. See associated
change in libcore/ that pulls it in, and see the change in
frameworks/base/ that uses it.

Bug: 118242715
Bug: 119026403
Test: build
Change-Id: I2bffce6b8ab469bf1a6d27300a90f210dd0784b1