Fix Morocco to use rearguard format correctly in ICU.
The previous change had no DST for these zones and changed the raw GMT
offset twice a year instead, which is problematic for the reasons
given in the comment in icu4c/source/tools/tzcode/icuzones in the
other change in this topic.

This change fixes test.java.time.format.TestZoneTextPrinterParser,
which was broken by the previous change (http://r.android.com/871773).

Bug: 122448923
Test: CtsLibcoreTestsCases
Test: CtsLibcoreOjTestCases (but see above)
Test: CtsIcuTestCases
Test: CtsBionicTestCases

Change-Id: I43744c3e147cfe4d7430ee86079e9792a1cff18c