Update out-of-date docs/fix download-iana-files.py
This commit updates the README.android to bring it up to date.
It also changes the flags on download-iana-files.py to make downloads
opt-in rather than opt-out, reflecting the common case is to only want
to download the data files.

Bug: 131159913
Test: Ran download-iana-files.py
Merged-In: Id1d8b66f8635a96a88591b19e64afb906bc9bd40
Change-Id: Id1d8b66f8635a96a88591b19e64afb906bc9bd40