Add a tool for generating a telephonylookup.xml
Add a tool for generating a telephonylookup.xml, which contains
telephony data related to time zone detection. This initial impl allows
the specification of MCC+MNC combinations that need special handling for
time zone detection because they are used outside of the expected
country.

The telephonylookup.xml is deliberately structured to allow for future
enhancements without needing to restructure the file.

This commit only contains the file generation code. The code on device
to read the file is separate, as is the changes to update-tzdata.py to
generate the real file, and the build changes needed to include the file
in the time zone data mainline module APEX.

Bug: 143451294
Test: atest telephonylookup_generator_test
Change-Id: I9c292ca9b816e0bb16932a473f9d2f324f76b7a9