Extend TzLookupGenerator to generate a new file
Extend TzLookupGenerator to generate a new "tz aliases" proto text
format file that contains information about alternative time zone IDs.

The intended usecase is to help canonicalize time zone IDs to the
ones used by Android, e.g. if we don't control which of several time
zone IDs an upstream data source uses, we can use the tz aliases to
identify the IDs that Android would use instead.

The tzaliases.txt file generated is not checked in as it's intended as
an input to a later stage of tz data generation that has yet to be
written. Instead, the file is written to an intermediates directory
(which is printed to stdout), e.g.
/tmp/tmpqXxOnA-tzdata/android_intermediates/tzaliases.txt
- this may change in a future commit.

Bug: 152746105
Test: Executed update-tzdata.py, inspected output
Test: atest --host input_tools/android/tzlookup_generator/src/test/
Change-Id: Ic03dfd523c841986c820fdd4c1d77b9ed5b53cf7