Tweak the tzdata format for future changes
The introduction of an explicit "final_offset" to the file will make it
easier to append sections later in a backwards-compatible way. It's
being added here also as a step towards removing the zone.tab section.

This commit also contains these tidy-up changes:
1. Tweaks to update the code to be a slightly more modern flavor of Java.

2. Modify the ZoneCompactor setup file format to prefix zone lines with
"Zone" rather than leaving the fact it's a reference to a zone file
implicit.

Bug: 161699306
Test: Ran update-tzdata.py to regenerate binary files
Test: treehugger
Test: Ran ZoneSplitter with dump-tzdata.py
Change-Id: I8d766fa24b88c9f9295a894dec05268f768a65df