Remove zone.tab from tzdata
The zone.tab information is not used on Android so can be
removed, saving 20KB in storage (and in Java heap).

NOTE TO OTHER ANDROID TIME ZONE MAINTAINERS:
This is not a backwards compatible change. tzdata files should
not be copied to earlier or later releases. Use update-tzdata.py from
the branch you are maintaining to create the binary files for that
branch.

Bug: 161699306
Test: update-tzdata.py
Test: Ran dump-tzdata.py
Change-Id: I6c79e9d2b9ef16679cacfb6ce4f011bfbb96faba