Extract tzaliases proto support code
Extract tzaliases proto support code, extend it and rename the proto to
extend the content. This change also makes the resulting tzids.prototxt
file from the update-tzdata.py pipeline an explicit output file in
output_data/android/. This is so it can be used in other data pipelines
that need to know what IDs Android recognizes, either globally or for a
specific country, and which zone IDs Android prefers.

Bug: 152747091
Test: atest --host input_tools/android/tzids/src/test/
Test: atest --host input_tools/android/tzlookup_generator/src/test/
Change-Id: Ib1af9db9d201c13c7ba70f88d870bae8187bc6d2