Fix TzIdProtos TimeZoneReplacement traversal logic
TimeZoneReplacement are a linked list. This updates the logic that
follows TimeZoneReplacement links to keep traversing until there are no
more replacements or the replacement is in the future. This also makes
the associated test include a list longer than one to prove that it
works.

A tz_ids_proto.proto comment was also fixed to correct an example:
America/Boise is the same as America/Denver and not America/Phoenix.

Bug: 171796385
Test: atest src/test/java/com/android/timezone/tzids/TimeZoneIdsTest.java
Change-Id: I54b3a6ed1a93540b0b3cd7da0f840fba6e26d564