Add mts-tzdata to test_suites
To support unbundled branches for mainline modules, we want to generate
partial MTS test suites that only contain testcases for a single module.
This converts converts all testcases for tzdata to also be included in
an android-mts-tzdata.zip.

If adding an additional _test module to MTS for tzdata, follow current
instructions, except adding "mts" to test_suites property, and add
`"mts-tzdata" to the test_suites property of that _test module. Having
an "mts-*" item in will also add the test to the overall mts test suite.

Test: m mts-tzdata and verify
  out/host/linux-x86/mts-tzdata/android-mts-tzdata/testcases/ contain
  MtsTimeZoneDataTestCases.apk MtsTimeZoneDataTestCases.config
Test: m mts and verify
  out/host/linux-x86/mts/android-mts/testcases/ contain
  MtsTimeZoneDataTestCases.apk MtsTimeZoneDataTestCases.config
Bug: 170318013
Change-Id: Ia1e83c3ad4c742b1b7164b0f7d1dc95e9b99a5b9