Do not use backward file in alternative time zones generation.
IANA's backward file has cross-country links and will have even more of
them in future releases as part of "merge time zone which are alike
since 1970".

This CL moves knowledge of alternative names to countryzones.txt file.
So it will be Android team's responsibility from now on to keep them
up-to-date and enable is to fix some existing issues with zones like
Africa/Asmera.

Bug: 190031717
Bug: 208768262
Test: ./update-tzdata.py generates the same tzlookup.xml
Test: fails when entire timeZoneMappings section in countryzones.txt is
commented out
Test: atest --host TzLookupGeneratorTest

Change-Id: I590a4980592ab348d4c093fce0e56fc3d68ed525