Make ZoneCompactor Links order agnostic.
Previously Link was not allowed to point to a time zone which
itself is just a Link to another time zone. Now it is allowed, and
order is not specified, so all Links have to be read in order to
tell whether a time zone is a genuine or a Link.

Bug: 254839538
Test: atest zone_compactor_tests
Test: ./update-tzdata.py changes no files
Change-Id: I3219aa131e10ab6f6b5ceeceb402de497607062d